Develop

+++ tests/helics/performance_tests/ring_tests.cpp
@@ -7,7 +7,7 @@ the top-level NOTICE for additional details. All rights reserved. See LICENSE fi
 #include <boost/test/unit_test.hpp>
 #include <boost/test/data/monomorphic.hpp>
 #include <boost/test/data/test_case.hpp>
-#include <boost/test/floating_point_comparison.hpp>
+#include <boost/test/tools/floating_point_comparison.hpp>
 
 #include "helics/application_api/Publications.hpp"
 #include "helics/application_api/Subscriptions.hpp"
@@ -145,7 +145,7 @@ BOOST_DATA_TEST_CASE (ring_test_multicores, bdata::make (fedCountB), feds)
         links[ii].initialize (cores[ii]->getIdentifier (), ii, feds);
     }
 
-    std::vector<std::thread> threads (feds + 1);
+    std::vector<std::thread> threads (static_cast<size_t>(feds) + 1);
     for (int ii = 0; ii < feds; ++ii)
     {
         threads[ii] = std::thread ([](RingTransmit &link) { link.run (); }, std::ref (links[ii]));
